Feature edit orders

+++ app/Http/Controllers/GroupController.php
@@ -26,6 +26,7 @@ class GroupController extends Controller
      * Display a listing of the resource.
      */
     public function index(Request $request)
+
     {
         // $this->authorize('viewAny',Group::class);
         $groups = $this->groupService->paginate($request);
@@ -33,12 +34,12 @@ class GroupController extends Controller
        
     }
 
-    /**
+        /**
      * Show the form for creating a new resource.
      */
     public function create()
     {
-        // $this->authorize('create',Group::class);
+        $this->authorize('create',Group::class);
         return view('admin.groups.create');
     }
 
@@ -47,6 +48,7 @@ class GroupController extends Controller
      */
     public function store(StoreGroupRequest $request)
     {
+      
         try {
             $this->groupService->store($request);
             toast('Thêm Quyền Thành Công!', 'success', 'top-right');
@@ -73,7 +75,7 @@ class GroupController extends Controller
     public function edit(string $id)
     {
         
-        // $this->authorize('update',Group::class);
+        $this->authorize('update',Group::class);
         $group = $this->groupService->find($id);
         return view('admin.groups.edit', compact('group') );
     }
@@ -100,6 +102,7 @@ class GroupController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Group::class);
         try {
             $this->groupService->forceDelete($id);
             toast('Nhóm Quyền Đã Được Xóa!', 'success', 'top-right');
